Reorganisation of imports to fix an import error
