+++ pyat/at/__init__.py
@@ -1,14 +1,6 @@
 """Python port of the Accelerator Toolbox"""
 
 # Make all functions visible in the at namespace:
+from .lattice import *
 from .tracking import *
 from .physics import *
-from .lattice import *
-
-
-class AtError(Exception):
-    pass
-
-
-class AtWarning(Warning):
-    pass
