+++ pyat/at/physics/__init__.py
@@ -7,5 +7,5 @@ from .matrix import *
 from .linear import *
 # noinspection PyUnresolvedReferences
 from .diffmatrix import find_mpole_raddiff_matrix
-from .radiation import *
 from .amat import *
+from .radiation import *
